# contracts/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpportunityDetailResponse": {
        "properties": {
          "baseline_selected": {
            "title": "Baseline Selected",
            "type": "boolean"
          },
          "calculated_at": {
            "title": "Calculated At",
            "type": "string"
          },
          "components": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Components",
            "type": "array"
          },
          "confidence_level": {
            "enum": [
              "high",
              "medium",
              "low",
              "not_publishable"
            ],
            "title": "Confidence Level",
            "type": "string"
          },
          "confidence_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "type": "number"
          },
          "definition_id": {
            "title": "Definition Id",
            "type": "string"
          },
          "definition_version": {
            "title": "Definition Version",
            "type": "integer"
          },
          "eligibility_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Eligibility Results",
            "type": "array"
          },
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "financial_scenario": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Financial Scenario"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "missing_features": {
            "items": {
              "type": "string"
            },
            "title": "Missing Features",
            "type": "array"
          },
          "property_unit_id": {
            "title": "Property Unit Id",
            "type": "string"
          },
          "publication_blockers": {
            "items": {
              "type": "string"
            },
            "title": "Publication Blockers",
            "type": "array"
          },
          "release_ids": {
            "items": {
              "type": "string"
            },
            "title": "Release Ids",
            "type": "array"
          },
          "score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "score_class": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "review",
                  "interesting",
                  "high_priority"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Class"
          },
          "segment_code": {
            "title": "Segment Code",
            "type": "string"
          },
          "snapshot_at": {
            "title": "Snapshot At",
            "type": "string"
          },
          "strategy": {
            "enum": [
              "division_extension",
              "renovation_resale"
            ],
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "id",
          "property_unit_id",
          "strategy",
          "score_class",
          "confidence_score",
          "confidence_level",
          "segment_code",
          "snapshot_at",
          "calculated_at",
          "baseline_selected",
          "definition_id",
          "definition_version",
          "eligible",
          "eligibility_results",
          "missing_features",
          "publication_blockers",
          "release_ids",
          "financial_scenario",
          "components"
        ],
        "title": "OpportunityDetailResponse",
        "type": "object"
      },
      "OpportunitySummaryResponse": {
        "properties": {
          "baseline_selected": {
            "title": "Baseline Selected",
            "type": "boolean"
          },
          "calculated_at": {
            "title": "Calculated At",
            "type": "string"
          },
          "confidence_level": {
            "enum": [
              "high",
              "medium",
              "low",
              "not_publishable"
            ],
            "title": "Confidence Level",
            "type": "string"
          },
          "confidence_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "property_unit_id": {
            "title": "Property Unit Id",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "score_class": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "review",
                  "interesting",
                  "high_priority"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Class"
          },
          "segment_code": {
            "title": "Segment Code",
            "type": "string"
          },
          "snapshot_at": {
            "title": "Snapshot At",
            "type": "string"
          },
          "strategy": {
            "enum": [
              "division_extension",
              "renovation_resale"
            ],
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "id",
          "property_unit_id",
          "strategy",
          "score_class",
          "confidence_score",
          "confidence_level",
          "segment_code",
          "snapshot_at",
          "calculated_at",
          "baseline_selected"
        ],
        "title": "OpportunitySummaryResponse",
        "type": "object"
      },
...
      "ScoreDefinitionResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "components": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Components",
            "type": "array"
          },
          "contract_digest": {
            "title": "Contract Digest",
            "type": "string"
          },
          "contract_path": {
            "title": "Contract Path",
            "type": "string"
          },
          "eligibility": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Eligibility",
            "type": "array"
          },
          "feature_registry_version": {
            "title": "Feature Registry Version",
            "type": "integer"
          },
          "features": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Features",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "meaning": {
            "title": "Meaning",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "publication_eligible": {
            "title": "Publication Eligible",
            "type": "boolean"
          },
          "strategy": {
            "enum": [
              "division_extension",
              "renovation_resale"
            ],
            "title": "Strategy",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "strategy",
          "contract_path",
          "contract_digest",
          "feature_registry_version",
          "parameters",
          "publication_eligible",
          "meaning",
          "active",
          "components",
          "features",
          "eligibility"
        ],
        "title": "ScoreDefinitionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Immo Opportunities API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/opportunities": {
      "get": {
        "operationId": "opportunities_api_v1_opportunities_get",
        "parameters": [
          {
            "in": "query",
            "name": "strategy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "division_extension",
                    "renovation_resale"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy"
            }
          },
          {
            "in": "query",
            "name": "minimum_score",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minimum Score"
            }
          },
          {
            "in": "query",
            "name": "confidence_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "high",
                    "medium",
                    "low",
                    "not_publishable"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Confidence Level"
            }
          },
          {
            "in": "query",
            "name": "department_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "22",
                    "29",
                    "35",
                    "56"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Department Code"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 250,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OpportunitySummaryResponse"
                  },
                  "title": "Response Opportunities Api V1 Opportunities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunities",
        "tags": [
          "scoring"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}": {
      "get": {
        "operationId": "opportunity_api_v1_opportunities__opportunity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity",
        "tags": [
          "scoring"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}/comparables": {
      "get": {
        "operationId": "opportunity_comparables_api_v1_opportunities__opportunity_id__comparables_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Opportunity Comparables Api V1 Opportunities  Opportunity Id  Comparables Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Comparables",
        "tags": [
          "scoring"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}/evidence": {
      "get": {
        "operationId": "opportunity_evidence_api_v1_opportunities__opportunity_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Opportunity Evidence Api V1 Opportunities  Opportunity Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Evidence",
        "tags": [
          "scoring"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}/history": {
      "get": {
        "operationId": "opportunity_history_api_v1_opportunities__opportunity_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OpportunitySummaryResponse"
                  },
                  "title": "Response Opportunity History Api V1 Opportunities  Opportunity Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity History",
        "tags": [
          "scoring"
        ]
      }
    },
...
    "/api/v1/opportunities/{opportunity_id}/sources": {
      "get": {
        "operationId": "opportunity_sources_api_v1_opportunities__opportunity_id__sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Opportunity Sources Api V1 Opportunities  Opportunity Id  Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Opportunity Sources",
        "tags": [
          "scoring"
        ]
      }
    },
...
    "/api/v1/score-definitions/{definition_id}": {
      "get": {
        "operationId": "score_definition_api_v1_score_definitions__definition_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "definition_id",
            "required": true,
            "schema": {
              "title": "Definition Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDefinitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score Definition",
        "tags": [
          "scoring"
        ]
      }
    },
...
  }
}